% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiword_terms}
\alias{multiword_terms}
\title{Multiword terms}
\format{A data frame with two variables: \code{term} and \code{sentiment}.}
\usage{
multiword_terms
}
\description{
(c) 2017, Wee Kim Wee School of Communication & Information,
Nanyang Technological University, Singapore
This WKWSCI Sentiment Lexicon ver 1.1 is licensed under the
Creative Commons Attribution-NonCommercial-ShareAlike 4.0
International License (http://creativecommons.org/licenses/by-nc-sa/4.0/)
}
\keyword{datasets}
