% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{with_pos}
\alias{with_pos}
\title{Lecixon with Part of Speech}
\format{A data frame with three variables: \code{term}, \code{POS}
and \code{sentiment}.}
\usage{
with_pos
}
\description{
(c) 2017, Wee Kim Wee School of Communication & Information,
Nanyang Technological University, Singapore
This WKWSCI Sentiment Lexicon ver 1.1 is licensed under the
Creative Commons Attribution-NonCommercial-ShareAlike 4.0
International License (http://creativecommons.org/licenses/by-nc-sa/4.0/)
}
\keyword{datasets}
