% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modifier}
\alias{modifier}
\title{Modifier lecixon}
\format{A data frame with three variables: \code{modifier}, \code{type}
and \code{POS}.}
\usage{
modifier
}
\description{
Lexicon contains modifiers such as absolutely and barely.
It is developed by Prof. Chris Khoo of
Wee Kim Wee School of Information and Communication, Nanyang
Technological University.
}
\keyword{datasets}
