% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiword_terms}
\alias{multiword_terms}
\title{Multiword terms}
\format{A data frame with two variables: \code{term} and \code{sentiment}.}
\usage{
multiword_terms
}
\description{
Lexicon that contains multiword terms, a.k.a short phrases.
It is developed by Prof. Chris Khoo of
Wee Kim Wee School of Information and Communication, Nanyang
Technological University.
}
\keyword{datasets}
