% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{with_pos}
\alias{with_pos}
\title{Lecixon with Part of Speech}
\format{A data frame with three variables: \code{term}, \code{POS}
and \code{sentiment}.}
\usage{
with_pos
}
\description{
Lexicon with part of speech is developed by Prof. Chris Khoo of
Wee Kim Wee School of Information and Communication, Nanyang
Technological University.
}
\keyword{datasets}
