% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{without_pos}
\alias{without_pos}
\title{Lecixon without Part of Speech}
\format{A data frame with two variables: \code{term} and \code{sentiment}.}
\usage{
without_pos
}
\description{
Lexicon without part of speech is developed by Prof. Chris Khoo of
Wee Kim Wee School of Information and Communication, Nanyang
Technological University.
}
\keyword{datasets}
